Python/Game/Snake-Royale/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.3.2
pydantic==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
dnspython==2.4.2  # Required for MongoDB Atlas SRV connection
pymongo[srv]==4.6.0